requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==0.61.1
django-cors-headers==4.5.0
django-environ==0.11.2
django-grappelli==4.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
urllib3==2.2.3
